调整视频编码渲染框架
