+++ EffectLib/src/main/java/com/gpufast/recoder/encoder/EncoderType.java
@@ -1,6 +1,6 @@
 package com.gpufast.recoder.encoder;
 
-public enum  EncoderType {
+public enum EncoderType {
 
     HW_VIDEO_ENCODER,
     SW_VIDEO_ENCODER,
