+++ EffectLib/src/main/java/com/gpufast/recoder/encoder/video/BaseBitrateAdjuster.java
@@ -1,12 +1,8 @@
 package com.gpufast.recoder.encoder.video;
 
-/**
- * BitrateAdjuster that tracks bitrate and framerate but does not adjust them.
- */
 class BaseBitrateAdjuster implements BitrateAdjuster {
     protected int targetBitrateBps;
     protected int targetFps;
-
     @Override
     public void setTargets(int targetBitrateBps, int targetFps) {
         this.targetBitrateBps = targetBitrateBps;
@@ -15,7 +11,7 @@ class BaseBitrateAdjuster implements BitrateAdjuster {
 
     @Override
     public void reportEncodedFrame(int size) {
-        // No op.
+
     }
 
     @Override
