+++ EffectLib/src/main/java/com/gpufast/recoder/encoder/video/MediaCodecWrapperFactory.java
@@ -1,5 +1,6 @@
 package com.gpufast.recoder.encoder.video;
 
+
 import java.io.IOException;
 
 interface MediaCodecWrapperFactory {
