+++ EffectLib/src/main/java/com/gpufast/recoder/encoder/video/VideoCodecStatus.java
@@ -17,7 +17,7 @@ public enum VideoCodecStatus {
 
   private final int number;
 
-  private VideoCodecStatus(int number) {
+   VideoCodecStatus(int number) {
     this.number = number;
   }
