--- EffectLib/src/main/java/com/gpufast/recoder/encoder/video2/BaseBitrateAdjuster.java
@@ -1,27 +1,0 @@
-package com.gpufast.recoder.encoder.video2;
-
-class BaseBitrateAdjuster implements BitrateAdjuster {
-    protected int targetBitrateBps;
-    protected int targetFps;
-
-    @Override
-    public void setTargets(int targetBitrateBps, int targetFps) {
-        this.targetBitrateBps = targetBitrateBps;
-        this.targetFps = targetFps;
-    }
-
-    @Override
-    public void reportEncodedFrame(int size) {
-
-    }
-
-    @Override
-    public int getAdjustedBitrateBps() {
-        return targetBitrateBps;
-    }
-
-    @Override
-    public int getCodecConfigFrameRate() {
-        return targetFps;
-    }
-}
