--- EffectLib/src/main/java/com/gpufast/recoder/encoder/video2/MediaCodecWrapperFactory.java
@@ -1,8 +1,0 @@
-package com.gpufast.recoder.encoder.video2;
-
-
-import java.io.IOException;
-
-interface MediaCodecWrapperFactory {
-  MediaCodecWrapper createByCodecName(String name) throws IOException;
-}
